fix(backend): split settings behavior and skip vendored codex in mypy
Move validators and derived properties into config_behavior.py so
config.py stays under the 500-line gate. Tell mypy to skip following
imports into the vendored openai_codex SDK so just typecheck stays green.
